updated the ui on register, login sidebar(admin)
