+++ student-registration-frontend/app/components/Sidebar.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { IconButton, Drawer, List, ListItem, ListItemIcon, ListItemText, Button, Box, Divider, ListItemButton } from '@mui/material';
+import { IconButton, Drawer, List, ListItem, ListItemIcon, ListItemText, Button, Box, Divider, ListItemButton, Typography, Avatar } from '@mui/material';
 import MenuIcon from '@mui/icons-material/Menu';
 import { FaUsers, FaTachometerAlt, FaSignOutAlt, FaClipboardList, FaUserEdit, FaCog } from 'react-icons/fa';
 import { useAuth } from '../context/AuthContext';
@@ -21,8 +21,8 @@ export default function Sidebar() {
 
   // Sidebar links for admin dashboard
   const sidebarLinks = [
-    { label: 'Dashboard', href: '/admin/dashboard', icon: <FaUsers /> },
-    { label: 'User Management', href: '/admin/dashboard#users', icon: <FaUserEdit /> },
+    { label: 'Dashboard', href: '/admin/dashboard', icon: <FaTachometerAlt /> },
+    { label: 'User Management', href: '/admin/users', icon: <FaUserEdit /> },
     { label: 'Audit Log', href: '/admin/audit-log', icon: <FaClipboardList /> },
     { label: 'Settings', href: '/admin/settings', icon: <FaCog /> },
   ];
@@ -67,7 +67,7 @@ export default function Sidebar() {
           display: { xs: 'none', sm: 'block' },
           '& .MuiDrawer-paper': {
             width: 260,
-            bgcolor: '#fff',
+            bgcolor: '#f8fafc',
             color: '#111',
             borderRight: '1px solid #e0e7ef',
             boxShadow: 2,
@@ -80,15 +80,15 @@ export default function Sidebar() {
         }}
         open
       >
-        <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, px: 3, py: 3, borderBottom: 1, borderColor: '#e0e7ef' }}>
-          <FaTachometerAlt style={{ color: '#6366f1', fontSize: 24 }} />
-          <span style={{ fontWeight: 'bold', fontSize: 20, color: '#111' }}>Admin Panel</span>
+        <Box sx={{ display: 'flex', alignItems: 'center', gap: 2, px: 3, py: 3, borderBottom: 1, borderColor: '#e0e7ef', bgcolor: '#6366f1' }}>
+          <Avatar sx={{ bgcolor: '#fff', color: '#6366f1', fontWeight: 700 }}>A</Avatar>
+          <Typography variant="h6" sx={{ fontWeight: 'bold', color: '#fff' }}>Admin Panel</Typography>
         </Box>
         <List sx={{ flex: 1, px: 1, py: 2 }}>
           {sidebarLinks.map(link => (
             <ListItem disablePadding key={link.label}>
-              <ListItemButton component="a" href={link.href} selected={pathname === link.href} sx={{ color: '#111', fontWeight: 600 }}>
-                <ListItemIcon sx={{ color: '#6366f1' }}>{link.icon}</ListItemIcon>
+              <ListItemButton component="a" href={link.href} selected={pathname === link.href} sx={{ color: pathname === link.href ? '#6366f1' : '#111', fontWeight: 600, borderRadius: 2, mb: 1, '&.Mui-selected': { bgcolor: '#e0e7ef' } }}>
+                <ListItemIcon sx={{ color: pathname === link.href ? '#6366f1' : '#6366f1' }}>{link.icon}</ListItemIcon>
                 <ListItemText primary={link.label} />
               </ListItemButton>
             </ListItem>
